pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Igor Churikov",email = "igorchrkv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standart] (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
